+++ rollup.config.mjs
@@ -3,6 +3,7 @@ import commonjs from '@rollup/plugin-commonjs'
 import esbuild from "rollup-plugin-esbuild"
 import dts from "rollup-plugin-dts"
 import alias from '@rollup/plugin-alias'
+import { string } from 'rollup-plugin-string'
 
 const preactAlias = alias({
   entries: [
@@ -18,11 +19,11 @@ const external = [
 ]
 
 export default [
-  ...createConfigs('src/entries/base.ts', 'dist/autocomplete.[ext]'),
-  ...createConfigs('src/entries/preact.ts', 'dist/preact/autocomplete.[ext]', preactAlias),
-  ...createConfigs('src/entries/react.ts', 'dist/react/autocomplete.[ext]'),
-  ...createConfigs('src/entries/liquid.ts', 'dist/liquid/autocomplete.[ext]'),
-  ...createConfigs('src/entries/mustache.ts', 'dist/mustache/autocomplete.[ext]'),
+  ...createConfigs('src/index.ts', 'dist/autocomplete.[ext]'),
+  ...createConfigs('src/react.ts', 'dist/preact/autocomplete.[ext]', preactAlias),
+  ...createConfigs('src/react.ts', 'dist/react/autocomplete.[ext]'),
+  ...createConfigs('src/liquid.ts', 'dist/liquid/autocomplete.[ext]'),
+  ...createConfigs('src/mustache.ts', 'dist/mustache/autocomplete.[ext]'),
 ]
 
 function createConfigs(input, outputTemplate, ...plugins) {
@@ -38,6 +39,9 @@ function createBuildConfig(input, outputTemplate, ...plugins) {
       resolve(), 
       commonjs(),
       esbuild(),
+      string({
+        include: '**/*.{mustache,liquid}',
+      }),
       ...plugins
     ],
     jsx: "react-jsx",
@@ -54,7 +58,10 @@ function createBuildConfig(input, outputTemplate, ...plugins) {
 function createDtsConfig(input, outputTemplate) {
   return {
     plugins: [
-      dts()
+      dts(),
+      string({
+        include: '**/*.{mustache,liquid}',
+      })
     ],
     input,
     output: {
