+++ spec/mustache.spec.ts
@@ -2,40 +2,13 @@ import { describe, expect, it } from "vitest"
 import "@testing-library/jest-dom"
 import {
   fromMustacheTemplate,
-  fromRemoteMustacheTemplate,
   defaultMustacheTemplate as mustacheTemplate,
-} from "../src/mustache"
+} from "../src/mustache/mustache"
 import {
-  autocompleteSuite,
-  handleAutocomplete,
-  hooks,
+  autocompleteSuite
 } from "./suites/autocomplete"
-import { waitFor } from "@testing-library/dom"
-import { mockFetch } from "./utils"
-
 describe("fromMustacheTemplate", () => {
   autocompleteSuite({
     render: fromMustacheTemplate(mustacheTemplate)
-  })
-})
-
-describe("fromRemoteMustacheTemplate", () => {
-  hooks()
-
-  it("fetches remote templates url", async () => {
-    const mockUrl = "template.mustache"
-    const render = fromRemoteMustacheTemplate(mockUrl)
-    mockFetch(mockUrl, mustacheTemplate)
-
-    await waitFor(() => handleAutocomplete(render))
-
-    await waitFor(
-      () => {
-        expect(global.fetch).toHaveBeenCalledWith(mockUrl)
-      },
-      {
-        timeout: 1000,
-      }
-    )
   })
 })
