+++ spec/suites/autocomplete.ts
@@ -8,8 +8,8 @@ import {
   AutocompleteConfig,
   DefaultState,
   autocomplete,
-} from "../../src/entries/base"
-import { getDefaultConfig } from "../../src/config"
+} from "../../src/index"
+import { getDefaultConfig } from "../../src/lib/config"
 import type { API, SearchResult } from "@nosto/nosto-js/client"
 import { mockNostojs } from "@nosto/nosto-js/testing"
