--- src/entries/base.ts
@@ -1,6 +1,0 @@
-/** @module ./ */
-export { autocomplete, type AutocompleteInstance, type SearchAutocompleteOptions } from "../autocomplete"
-export { search } from "../search"
-export type { AutocompleteConfig, GoogleAnalyticsConfig, Selector } from "../config"
-export { type DefaultState } from "../utils/state"
-export { priceDecorator } from "@nosto/search-js/currencies"
