--- src/entries/liquid.ts
@@ -1,12 +1,0 @@
-/** @module liquid */
-export {
-  fromLiquidTemplate,
-  fromRemoteLiquidTemplate,
-  defaultLiquidTemplate,
-} from "../liquid"
-export {
-  autocomplete, type AutocompleteInstance, search, 
-  type AutocompleteConfig, type GoogleAnalyticsConfig,
-  type DefaultState,
-  priceDecorator
-} from "./base"
