--- src/entries/mustache.ts
@@ -1,13 +1,0 @@
-/** @module mustache */
-export {
-  fromMustacheTemplate,
-  fromRemoteMustacheTemplate,
-  defaultMustacheTemplate,
-  type Options
-} from "../mustache"
-export {
-  autocomplete, type AutocompleteInstance, search, 
-  type AutocompleteConfig, type GoogleAnalyticsConfig,
-  type DefaultState,
-  priceDecorator
-} from "./base"
