--- src/entries/preact.ts
@@ -1,7 +1,0 @@
-export { Autocomplete, type AutocompleteProps } from "../defaults/Autocomplete"
-export {
-    autocomplete, type AutocompleteInstance, search, 
-    type AutocompleteConfig, type GoogleAnalyticsConfig,
-    type DefaultState,
-    priceDecorator
-  } from "./base"
