+++ src/lib/autocomplete.ts
@@ -1,12 +1,12 @@
-import { logger, recordSearchClick, recordSearchSubmit } from "./api/client"
+import { logger, recordSearchClick, recordSearchSubmit } from "./client"
 import { AutocompleteConfig, getDefaultConfig } from "./config"
-import { Dropdown, createDropdown, parseHit } from "./utils/dropdown"
-import { DefaultState, StateActions, getStateActions } from "./utils/state"
-import { bindClickOutside, findAll } from "./utils/dom"
-import { bindInput } from "./utils/input"
-import { getGaTrackUrl, isGaEnabled, trackGaPageView } from "./utils/ga"
-import createDebouncer from "./utils/debounce"
-import { createHistory } from "./utils/history"
+import { Dropdown, createDropdown, parseHit } from "./dropdown"
+import { DefaultState, StateActions, getStateActions } from "../utils/state"
+import { bindClickOutside, findAll } from "../utils/dom"
+import { bindInput } from "../utils/input"
+import { getGaTrackUrl, isGaEnabled, trackGaPageView } from "../utils/ga"
+import createDebouncer from "../utils/debounce"
+import { createHistory } from "../utils/history"
 import type { SearchOptions } from "@nosto/nosto-js/client"
 
 export type AutocompleteInstance = {
