+++ src/liquid/liquid.ts
@@ -1,0 +1,57 @@
+import { DefaultState } from "../utils/state"
+import { Liquid } from "liquidjs"
+
+export { default as defaultLiquidTemplate } from "./autocomplete.liquid"
+
+/**
+ * Render a liquid template into a container
+ *
+ * @param template Liquid template
+ * @returns Render function
+ * @group Autocomplete
+ * @category Liquid
+ * @example
+ * ```js
+ * import { fromLiquidTemplate } from "@nosto/autocomplete/liquid";
+ *
+ * const render = fromLiquidTemplate(`
+ *   <div>
+ *      <h1>{{title}}</h1>
+ *    <ul>
+ *    {% for product in products %}
+ *     <li>{{product.name}}</li>
+ *  {% endfor %}
+ *  </ul>
+ *  </div>
+ *  `);
+ *
+ * render(document.getElementById("container"), {
+ *   title: "My Title",
+ *   products: [
+ *     { name: "Product 1" },
+ *     { name: "Product 2" },
+ *     { name: "Product 3" }
+ *   ]
+ * });
+
+ * ```
+ */
+export function fromLiquidTemplate<State extends object = DefaultState>(
+  template: string
+): (container: HTMLElement, state: State) => PromiseLike<void> {
+  const instance = Liquid ? new Liquid() : undefined
+
+  if (instance === undefined) {
+    throw new Error(
+      "Liquid is not defined. Please include the Liquid library in your page."
+    )
+  }
+
+  const tmpl = instance.parse(template)
+
+  return (container, state) => {
+    container.innerHTML = instance.renderSync(tmpl, state)
+
+    return Promise.resolve(undefined)
+  }
+}
