+++ src/react.ts
@@ -1,8 +1,8 @@
 /** @module react */
-export { Autocomplete, type AutocompleteProps } from "../defaults/Autocomplete"
+export { Autocomplete, type AutocompleteProps } from "./react/Autocomplete"
 export {
     autocomplete, type AutocompleteInstance, search, 
     type AutocompleteConfig, type GoogleAnalyticsConfig,
     type DefaultState,
     priceDecorator
-  } from "./base"
+  } from "."
