+++ vitest.config.ts
@@ -1,6 +1,19 @@
 import { defineConfig } from "vite"
 
 export default defineConfig({
+  plugins: [
+    {
+      name: 'transform-mustache-liquid',
+      transform(code, id) {
+        if (/\.(mustache|liquid)$/.test(id)) {
+          return {
+            code: `export default ${JSON.stringify(code)};`,
+            map: null,
+          };
+        }
+      },
+    }
+  ],
   test: {
     include: ["./spec/**/*.spec.{ts,tsx}"],
     environment: "jsdom",
